--- a/src/common/types/types.cpp
+++ b/src/common/types/types.cpp
@@ -232,9 +232,9 @@ std::string PhysicalTypeUtils::toString(PhysicalTypeID physicalType) {
         return "ARRAY";
     case PhysicalTypeID::POINTER:
         return "POINTER";
-    case kuzu::common::PhysicalTypeID::ALP_EXCEPTION_FLOAT:
+    case PhysicalTypeID::ALP_EXCEPTION_FLOAT:
         return "ALP_EXCEPTION_FLOAT";
-    case kuzu::common::PhysicalTypeID::ALP_EXCEPTION_DOUBLE:
+    case PhysicalTypeID::ALP_EXCEPTION_DOUBLE:
         return "ALP_EXCEPTION_DOUBLE";
     default:
         KU_UNREACHABLE;
@@ -272,10 +272,10 @@ uint32_t PhysicalTypeUtils::getFixedTypeSize(PhysicalTypeID physicalType) {
         return sizeof(interval_t);
     case PhysicalTypeID::INTERNAL_ID:
         return sizeof(internalID_t);
-    case kuzu::common::PhysicalTypeID::ALP_EXCEPTION_FLOAT:
-        return storage::EncodeException<float>::sizeBytes();
-    case kuzu::common::PhysicalTypeID::ALP_EXCEPTION_DOUBLE:
-        return storage::EncodeException<double>::sizeBytes();
+    case PhysicalTypeID::ALP_EXCEPTION_FLOAT:
+        return storage::EncodeException<float>::sizeInBytes();
+    case PhysicalTypeID::ALP_EXCEPTION_DOUBLE:
+        return storage::EncodeException<double>::sizeInBytes();
     default:
         KU_UNREACHABLE;
     }