--- a/src/include/common/utils.h
+++ b/src/include/common/utils.h
@@ -43,5 +43,26 @@ bool isLittleEndian();
 template<typename T>
 bool integerFitsIn(int64_t val);
 
+template<typename T>
+std::vector<T> copyVector(const std::vector<T>& objects) {
+    std::vector<T> result;
+    result.reserve(objects.size());
+    for (auto& object : objects) {
+        result.push_back(object->copy());
+    }
+    return result;
+}
+
+template<numeric_utils::IsIntegral T>
+constexpr T ceilDiv(T a, T b) {
+    return (a / b) + (a % b != 0);
+}
+
+template<std::integral To, std::integral From>
+constexpr To safeIntegerConversion(From val) {
+    KU_ASSERT(static_cast<To>(val) == val);
+    return val;
+}
+
 } // namespace common
 } // namespace kuzu