--- a/src/include/storage/store/column.h
+++ b/src/include/storage/store/column.h
@@ -3,10 +3,9 @@
 #include "catalog/catalog.h"
 #include "common/null_mask.h"
 #include "common/types/types.h"
-#include "storage/compression/compression.h"
 #include "storage/db_file_id.h"
 #include "storage/store/column_chunk_data.h"
-#include "storage/store/column_read_writer.h"
+#include "storage/store/column_reader_writer.h"
 
 namespace kuzu {
 namespace evaluator {