--- a/src/include/storage/store/column_chunk.h
+++ b/src/include/storage/store/column_chunk.h
@@ -42,7 +42,7 @@ public:
         ColumnChunkMetadata metadata);
     ColumnChunk(bool enableCompression, std::unique_ptr<ColumnChunkData> data);
 
-    void initializeScanState(ChunkState& state) const;
+    void initializeScanState(ChunkState& state, Column* column) const;
     void scan(const transaction::Transaction* transaction, const ChunkState& state,
         common::ValueVector& nodeID, common::ValueVector& output, common::offset_t offsetInChunk,
         common::length_t length) const;