# Description

Please include a summary of the changes and the related issue (if applicable). Please also include
relevant motivation and context.

Fixes # (issue)

# Contributor agreement

- [ ] I have read and agree to the [Contributor Agreement](https://github.com/kuzudb/kuzu/blob/master/CLA.md).