@royi-luo 2024-07-31
Some older benchmarks can be found in [this PR](https://github.com/kuzudb/kuzu/pull/3712)