@royi-luo 2024-08-03
Benchmarks for query times + compression ratio

# Compression Ratio (Sizes are in KB)

**Baseline Compression Ratios for DuckDB can be found [here](https://github.com/kuzudb/kuzu/pull/3712#issuecomment-2229050188)**

Dataset|Database Size (master)|Database Size (ALP)|Compression Ratio
----|----|----|----
Medicare2|895707.2|613256.0|1.46
CommonGovernment|1508672.0|640703.6|2.35x

# Copy Time (in ms)
Dataset|Copy Time (master)|Copy Time (ALP)
----|----|----
Medicare2|2748.727|2564.315
CommonGovernment|4883.586|5315.646

# Query Time (in ms)
Query: `MATCH (row:{Table}) RETURN MIN(row.ID), MIN(row.{Column1}), MIN(row.{Column2}), ...`
Dataset|Query Time (master)|Query Time (ALP)
----|----|----
Medicare2|84.939|106.09
CommonGovernment|165.00300000000001|131.29


# Update Times (in ms)
## No New ALP Exceptions
Query: 
```
CALL auto_checkpoint=false;
MATCH (row:{Table}) SET row.{Column} = {non-exception-value};
CHECKPOINT;
```
Dataset|Update Time (master)|Checkpoint Time(master)|Update Time (ALP)|Checkpoint Time (ALP)
----|----|----|----|----
Medicare2|9722.37|479.786|10058.74|759.866
CommonGovernment|13941.887|642.949|14606.534|1158.7410000000002

## 10% of new values are ALP exceptions
Query: 
```
CALL auto_checkpoint=false;
MATCH (row:{Table}) SET row.{Column} = 
CASE
  WHEN row.ID % 10 = 0 THEN {non-exception-value}
  ELSE 1000000.123456789
END;
CHECKPOINT;
```
Dataset|Update Time (master)|Checkpoint Time(master)|Update Time (ALP)|Checkpoint Time (ALP)
----|----|----|----|----
Medicare2|11147.333|480.653|12092.397|673.787
CommonGovernment|13941.887|642.949|14606.534|1158.7410000000002